+++ R/subset_network.R
@@ -46,16 +46,25 @@ subset_network <- function(x, outlet, only_up = FALSE) {
 #' @name subset_network
 #' @export
 subset_network.data.frame <- function(x, outlet, only_up = FALSE) {
-  x <- hy(x)
-
-  x <- subset_network(x, outlet, only_up)
-
-  hy_reverse(x)
+  hy_as_dataframe(x, "subset_network", outlet = outlet, only_up = only_up)
 }
 
 #' @name subset_network
 #' @export
 subset_network.hy <- function(x, outlet, only_up = FALSE) {
+  hy_classify_and_redispatch(x, "subset_network", "hy_node", hy_guidance_node,
+    outlet = outlet, only_up = only_up)
+}
+
+#' @name subset_network
+#' @export
+subset_network.hy_topo <- function(x, outlet, only_up = FALSE) {
+  hy_topo_to_node(x, "subset_network", outlet = outlet, only_up = only_up)
+}
+
+#' @name subset_network
+#' @export
+subset_network.hy_node <- function(x, outlet, only_up = FALSE) {
 
   check_names(x, c(id, fromnode, tonode), "subset_network")
 
@@ -91,7 +100,7 @@ subset_network.hy <- function(x, outlet, only_up = FALSE) {
 
       down <- navigate_network_dfs(index_ids$to, missed, "down")
 
-      outlets <- x$id[x$toid == get_outlet_value(x) & x$id %in% unlist(down)]
+      outlets <- x$id[is_outlet(x) & x$id %in% unlist(down)]
 
       more_up <- unlist(navigate_network_dfs(index_ids$from, outlets, "up"))